+++ Assets/CommonCoreModules/World/ObjectActions/OnTriggerEnterTrigger.cs
@@ -17,6 +17,7 @@ namespace CommonCore.ObjectActions
         public bool OnActorsOnly = false;
 
         public bool CheckAllCollisions = false;
+        public bool CheckParentsForEntity = true;
 
         void Start()
         {
@@ -29,9 +30,17 @@ namespace CommonCore.ObjectActions
                 return;
 
             var otherGameObject = other.gameObject;
-            var entity = other.gameObject.GetEntity();
-            if (entity != null)
-                otherGameObject = entity.gameObject;
+            BaseController entity;
+            if (CheckParentsForEntity)
+            {
+                entity = other.gameObject.GetEntity();
+                if (entity != null)
+                    otherGameObject = entity.gameObject;
+            }            
+            else
+            {
+                entity = otherGameObject.GetComponent<BaseController>();
+            }
 
             //reject not-player if we're not allowing not-player
             if (OnPlayerOnly && !otherGameObject.IsPlayer())
@@ -42,8 +51,7 @@ namespace CommonCore.ObjectActions
                 return;
 
             //execute special
-            var activator = entity;
-            var data = new ActionInvokerData() { Activator = activator, Caller = this };
+            var data = new ActionInvokerData() { Activator = entity, Caller = this };
             Special.Invoke(data);
 
             //lock if not repeatable